refactor(services): Remove duplication on date operations

+++ src/main/java/fr/caprog/tdd/common/DayShift.java
@@ -1,7 +1,7 @@
 package fr.caprog.tdd.common;
 
 public class DayShift {
-    public static final int ONE_DAY = 1;
-    public static final int TWO_DAYS = 2;
-    public static final int THREE_DAYS = 3;
+    public static final long ONE_DAY = 1;
+    public static final long TWO_DAYS = 2;
+    public static final long THREE_DAYS = 3;
 }
